requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
ollama
faiss-cpu
sentence-transformers
pypdf
sqlalchemy
pydantic
python-dotenv